Minimal changes for production usage (#11)
"minimal" describes the kind of changes: They are just the bare minimum
required to start it in the production setup and only set the admin
password to either the secret given or the default "admin".

I've added a entrypoint to wait for dependend services. Otherwise I got
errors, that e.g. the datastore is not reachable.

I've added a entrypoint-setup which is used as a one-time container: In
it, the necessary manage commands are executed when the setup starts.
More things to do will follow (like configurating the organization and
so on). This must be done in a docker context to have this automatically
run. The way we run our prod setup is that it must be self-contained -
it is configured through env/secrets and after a deployment it must be
ready. This is why I execute ./manage in the dockercontainer itself.

More changes:
- The service now has the preallocated port 9008
- The EXEC format in the dockerfile has not the recommended format

Some changes needs to be cleaned up later:
- All print/log statements are missing \n
- The server must not listen on `MANAGE_HOST`. If so, no other service
can reach the container. I just removed the host form the address, maybe
this needs to be refactored later.

And yes, I used bash. Because it gets the job done and this is a part,
where also our server maintainers will change stuff - and bash is a
commonly known tool which can be easily scripted.

I'll merge this fast since we need this to have our prod setup running
for our client. I'll write an email with additionally needs that can be
implemented later on.